task_03/requirements.txt
Flask
python-dotenv
Werkzeug
bandit
flake8
safety
